Split `arg-input` into a folder of focused control files
